+++ pypgx/api/pipeline.py
@@ -13,7 +13,7 @@ from . import utils, plot, genotype, core
 def run_ngs_pipeline(
     gene, output, variants=None, depth_of_coverage=None,
     control_statistics=None, panel=None,
-    force=False, do_not_plot_copy_number=False,
+    force=False, samples=None, do_not_plot_copy_number=False,
     do_not_plot_allele_fraction=False
 ):
     """
@@ -36,6 +36,8 @@ def run_ngs_pipeline(
         Reference haplotype panel.
     force : bool, default : False
         Overwrite output directory if it already exists.
+    samples : list, optional
+        List of known samples without SV.
     do_not_plot_copy_number : bool, default: False
         Do not plot copy number profile.
     do_not_plot_allele_fraction : bool, default: False
@@ -88,7 +90,7 @@ def run_ngs_pipeline(
 
         read_depth = utils.import_read_depth(gene, depth_of_coverage)
         read_depth.to_file(f'{output}/read-depth.zip')
-        copy_number = utils.compute_copy_number(read_depth, control_statistics)
+        copy_number = utils.compute_copy_number(read_depth, control_statistics, samples=samples)
         copy_number.to_file(f'{output}/copy-number.zip')
         cnv_calls = utils.predict_cnv(copy_number)
         cnv_calls.to_file(f'{output}/cnv-calls.zip')