+++ pypgx/cli/run_ngs_pipeline.py
@@ -9,6 +9,8 @@ description = f"""
 # Run NGS pipeline for the target gene. #
 #########################################
 
+When computing copy number from read depth, if the input data was generated with targeted sequencing as opposed to WGS, the method will apply inter-sample normalization using summary statistics across all samples. For best results, it is recommended to manually specify a list of known reference samples that do not have SV with '--samples'.
+
 Usage examples:
   $ pypgx {fuc.api.common._script_name()} CYP2D6 CYP2D6-pipeline --variants variants.vcf --depth-of-coverage depth-of-coverage.tsv --control-statistcs control-statistics-VDR.zip
 """
@@ -54,6 +56,12 @@ def create_parser(subparsers):
         help='Overwrite output directory if it already exists.'
     )
     parser.add_argument(
+        '--samples',
+        metavar='TEXT',
+        nargs='+',
+        help='List of known samples with no SV.'
+    )
+    parser.add_argument(
         '--do-not-plot-copy-number',
         action='store_true',
         help='Do not plot copy number profile.'
@@ -69,7 +77,7 @@ def main(args):
         args.gene, args.output, variants=args.variants,
         depth_of_coverage=args.depth_of_coverage,
         control_statistics=args.control_statistics,
-        panel=args.panel, force=args.force,
+        panel=args.panel, force=args.force, samples=args.samples,
         do_not_plot_copy_number=args.do_not_plot_copy_number,
         do_not_plot_allele_fraction=args.do_not_plot_allele_fraction
     )