+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * Update :meth:`api.utils.sort_alleles` method to give priority to alleles that impact protein coding when breaking ties (i.e. alleles have the same functional status and same number of variants).
 * Update CNV caller for SLC22A2 and UGT2B15 genes.
 * Rename ``--chr-prefix`` argument in :command:`create-regions-bed` to ``--add-chr-prefix``.
+* Add ``--samples`` argument to :command:`run-ngs-pipeline` command.
 
 0.6.0 (2021-10-09)
 ------------------